requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
dj-database-url==0.5.0
Django==4.0.4
django-cors-headers==3.11.0
django-shortcuts==1.6
djangorestframework==3.13.1
flake8==4.0.1
gunicorn==20.1.0
mccabe==0.6.1
psycopg2==2.9.3
psycopg2-binary
pycodestyle==2.8.0
pyflakes==2.4.0
python-decouple==3.6
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
whitenoise==6.0.0
